Added buttons to reset scores and restart level
Added a reset button underneath the highscore list that resets the saved high scores.

Added a restart level button that shows up when the player(s) died.

+++ Lightweight_2D/GameObject.cpp
@@ -11,7 +11,7 @@ LW2D::GameObject::GameObject(std::string name)
 
 void LW2D::GameObject::Initialize()
 {
-	for (int i = 0; i < m_pComponents.size(); i++)
+	for (int i = 0; i < static_cast<int>(m_pComponents.size()); i++)
 	{
 		m_pComponents[i]->Initialize();
 	}
