+++ Lightweight_2D/SceneManager.cpp
@@ -4,7 +4,7 @@
 
 void LW2D::SceneManager::Initialize()
 {
-	for (int i = 0; i < m_scenes.size(); i++)
+	for (int i = 0; i < static_cast<int>(m_scenes.size()); i++)
 	{
 		m_ActiveSceneIndex = static_cast<uint8_t>(i);
 		m_scenes[i]->Initialize();
